Rename logging package to extensions
- The logging package only had a single class: The logging extension. Because we'll be introducing more extensions in the future, the package was renamed to extensions to group them all together in there.
