+++ utilities/test-util/src/main/java/nl/pim16aap2/testing/extensions/package-info.java
@@ -1,0 +1,8 @@
+/**
+ * This package contains a set of {@link org.junit.jupiter.api.extension.Extension}s.
+ * <p>
+ */
+@NonNullByDefault
+package nl.pim16aap2.testing.extensions;
+
+import org.eclipse.jdt.annotation.NonNullByDefault;
